requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose
passlib[bcrypt]